Clarify hotswap and switches for the Dilemma.
I started my first Dilemma build, and was a bit confused by some things in the docs.

First, the docs specifically call out 3D keyboards not being compatible with hotswap,
but say nothing about the Dilemma. After asking on the Discord, I found out that the
Dilemma also does not support hotswap, which I think covers all keyboards from BastardKB,
so I don't think the "3d keyboards" qualification is necessary.

Second, I couldn't find any information about what kind of switches are supported on either
the docs or the product page. It would be great to add this to the product page as well,
where it says "Keycaps, switches and cables not included.".

I got the information on compatibility from `burkfers` on Discord.